Add tests to new classes
